updated test imports to reflect tree relocation
